+++ R/boilerplate_methods_eligibility_criteria.R
@@ -6,7 +6,7 @@
 #'
 #' @return A character string containing the eligibility criteria description in markdown format.
 #'
-#' @keywords internal
+#' @export
 boilerplate_methods_eligibility_criteria <- function(inclusion_criteria, exclusion_criteria, n_participants, baseline_wave) {
   # Function to convert a list to markdown bullet points
   list_to_markdown <- function(criteria_list) {