+++ R/boilerplate_methods_target_population.R
@@ -4,7 +4,7 @@
 #'
 #' @return A character string containing the target population description in markdown format.
 #'
-#' @keywords internal
+#' @export
 boilerplate_methods_target_population <- function(statistical_estimator = "lmtp", baseline_wave) {
   base_text <- glue::glue("## Target Population
 The target population for this study comprises New Zealand residents as represented in the {baseline_wave} of the New Zealand Attitudes and Values Study (NZAVS) during the years {baseline_wave} weighted by New Zealand Census weights for age, gender, and ethnicity (refer to @sibley2021). The NZAVS is a national probability study designed to reflect the broader New Zealand population accurately. Despite its comprehensive scope, the NZAVS has some limitations in its demographic representation. Notably, it tends to under-sample males and individuals of Asian descent while over-sampling females and Māori (the indigenous peoples of New Zealand). To address these disparities and enhance the accuracy of our findings, we apply New Zealand Census survey weights to the sample data.")